requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
decorator==4.4.2
Django==4.2.5
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.4
imageio==2.31.4
imageio-ffmpeg==0.4.9
moviepy==1.0.3
numpy==1.26.0
packaging==23.1
Pillow==10.0.1
proglog==0.1.10
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
SpeechRecognition==3.10.0
sqlparse==0.4.4
tqdm==4.66.1
tzdata==2023.3
urllib3==2.0.5
